fix(migration): whatsapp campaign person_id type matches persons.id
persons.id en tenants legacy es `int unsigned`, no `bigint unsigned`.
La migración creaba whatsapp_offer_campaign_messages.person_id como
unsignedBigInteger, que hacía fallar la FK con MySQL error 3780:
"Referencing column and referenced column are incompatible".

El bloqueo impedía que cualquier migración posterior corriera — incluidas
las 3 del flujo de pedidos (2026_04_20_*).

La migración es idempotente (usa Schema::hasTable), así que tenants que
ya tengan la primera tabla solo crearán la segunda con tipo correcto.

+++ database/migrations/tenant/2026_04_15_000001_create_whatsapp_offer_campaign_tables.php
@@ -31,7 +31,9 @@ class CreateWhatsappOfferCampaignTables extends Migration
             Schema::create('whatsapp_offer_campaign_messages', function (Blueprint $table) {
                 $table->id();
                 $table->unsignedBigInteger('campaign_id');
-                $table->unsignedBigInteger('person_id');
+                // persons.id es `int unsigned` (legacy), NO bigint — debe coincidir
+                // o MySQL rechaza la FK con error 3780 (incompatible columns).
+                $table->unsignedInteger('person_id');
                 $table->string('phone', 30)->nullable();
                 $table->string('status', 20)->default('pending'); // pending|sent|failed|skipped
                 $table->json('payload')->nullable();
